requirements.txt
flet==0.83.0
mysql-connector-python
bcrypt